+++ src/main/java/me/chaseoes/tf2/SpectatePlayer.java
@@ -39,7 +39,7 @@ public class SpectatePlayer {
         savedXPCount = player.getExp();
         savedLevelCount = player.getLevel();
         savedFoodLevel = player.getFoodLevel();
-        savedHealth = player.getHealth();
+        savedHealth = (int) player.getHealth();
         savedGameMode = player.getGameMode();
     }
